Revert "x64 Debug でのコンパイルエラーを一時的に解消するために capacity() の戻り値を戻す"
This reverts commit a2bb4797f629825ff13b34ca3069873500a8ff68.

+++ sakura_core/mem/CNativeW.h
@@ -135,7 +135,7 @@ public:
 	void swap( CNativeW& left ){
 		CNative::swap(left);
 	}
-	int capacity() const noexcept {
+	SSIZE_T capacity() const noexcept {
 		return CNative::capacity() / sizeof(wchar_t);
 	}
